requirements.txt
python-docx~=0.8.10
environs~=9.3.2
openpyxl~=3.0.7